requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchainhub>=0.1.15
langgraph>=0.0.40
llama-cpp-python>=0.2.70
sentence-transformers>=2.6.1
transformers>=4.41.0
huggingface-hub>=0.22.2
qdrant-client>=1.8.0
duckduckgo-search>=4.3
python-dotenv>=1.0.1
tqdm>=4.66.4
numpy>=1.26.4
pypdf>=3.17.4
uuid>=1.30
streamlit>=1.35.0
rich>=13.7.1
